Style clean up and renames
Renamed files to match naming conventions. Same with function names.
Note that there is still a style error for having the path command in
front of the import(s), but that is beyond my control at this point. If
the append is not there first, then the import won't find the files.
That will be fixed when I make this into a package.